requirements.txt
# ============================================================
# Core dependencies (required for all setups)
# ============================================================
chromadb>=0.5.0
sentence-transformers>=3.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0

# Document parsing
pymupdf>=1.24.0       # PDF  (import as 'fitz' or 'pymupdf')
python-docx>=1.1.0    # DOCX

# ============================================================
# Cloud provider SDKs — install only what you need
# ============================================================
# pip install openai          # OpenAI (GPT-4o, etc.)
# pip install anthropic       # Claude (Anthropic)
# pip install mistralai       # Mistral AI
